+++ CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Public command:
     - Get-RegistryPropertyValue - Get a registry property value.
 
+### Changed
+
+- `Compare-DscParameterState`
+  - Updated comment-based help example.
+
 ## [0.19.1] - 2025-02-08
 
 ### Added